renamed getStoragedAccount to getLoggedInAccount
