+++ src/background/utils/getLoggedInAccount.ts
